Formaat Refactor
Refactored the project into a server client system where images and comments are hosted on the server and the client has the images and creates the butytons and fetches the dataa from the server. Server not yet implemented. Client now creates buttons of all included images. Created scripts to copy images and comments to required locations and injest images from a folder and create directoris with the image and comments in the central repo.